pyproject.toml
[tool.poetry]
name = "kingsceo"
version = "0.1.0"
description = "Multi-bot Telegram orchestration platform"
authors = ["Kings CEO Team <dev@kingsceo.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.7"
SQLAlchemy = "^2.0.33"
asyncpg = "^0.29.0"
alembic = "^1.14.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
structlog = "^24.1.0"
cachetools = "^5.3.3"
aiohttp = "^3.10.5"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
cryptography = "^43.0.0"
uvloop = {version = "^0.20.0", markers = "sys_platform != 'win32'"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
ruff = "^0.6.9"
mypy = "^1.11.2"
types-cachetools = "^5.3.0.6"

[tool.poetry.scripts]
kingsceo-bot = "app.bots.runner:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
